+++ archetype-core/src/main/java/org/apiaddicts/apitools/apigen/archetypecore/core/advice/ApigenControllerAdvice.java
@@ -25,6 +25,7 @@ import org.springframework.web.servlet.NoHandlerFoundException;
 
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
+
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Parameter;