+++ archetype-core/src/test/java/org/apiaddicts/apitools/apigen/archetypecore/core/advice/ApigenControllerAdviceTests.java
@@ -2,8 +2,8 @@ package org.apiaddicts.apitools.apigen.archetypecore.core.advice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.web.servlet.MockMvc;