+++ generator-cli/src/main/java/org/apiaddicts/apitools/apigen/generatorcli/App.java
@@ -1,0 +1,13 @@
+package org.apiaddicts.apitools.apigen.generatorcli;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.shell.command.annotation.CommandScan;
+
+@CommandScan
+@SpringBootApplication
+public class App {
+    public static void main(String[] args) {
+        SpringApplication.run(App.class, args);
+    }
+}