+++ generator-cli/src/main/java/org/apiaddicts/apitools/apigen/generatorcli/config/AutoconfigureExclusionsConfig.java
@@ -1,12 +1,16 @@
-package org.apiaddicts.apitools.apigen.generatorrest.core.config;
+package org.apiaddicts.apitools.apigen.generatorcli.config;
 
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.hibernate.autoconfigure.HibernateJpaAutoConfiguration;
 import org.springframework.boot.jdbc.autoconfigure.DataSourceAutoConfiguration;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @EnableAutoConfiguration(exclude = {HibernateJpaAutoConfiguration.class, DataSourceAutoConfiguration.class})
 public class AutoconfigureExclusionsConfig {
-	// Intentional blank
+    @Bean
+    CustomExceptionResolver customExceptionResolver() {
+        return new CustomExceptionResolver();
+    }
 }